backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
google-cloud-aiplatform>=1.38.0
vertexai>=1.43.0
pydantic>=2.9.0
requests>=2.31.0
langchain>=0.1.0
langchain-core>=0.1.0
python-dotenv>=1.0.0

